#158 Re-did the layout options, reduced to top, left, right and bottom
Whole layout is now based on (mostly) grid, some flex used as well. Layout is now more pure, and consistent. Added a lot of examples, and ensured that all options work similar

 - Setting width by setting width of container div
 - Setting width by configuration parameter
 - Setting boardSize by configuration parameter

Some changes where done to the headers / tags, the colorMarker for who should move next, the fen. Edit mode is consistent as well, but not finished.